完善后端系统功能开发

+++ dts-dao/src/main/java/com/qiguliuxing/dts/vo/RewardRulesVo.java
@@ -1,12 +1,13 @@
 package com.qiguliuxing.dts.vo;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class RewardRulesVo {
     private Integer id;          // 主键ID
     private String rankName;     // 名次(如第一名、第二名)
     private Integer rankOrder;   // 排名顺序(1,2,3...)
-    private Double percentage;   // 奖励百分比(如35.00表示35%)
+    private BigDecimal percentage;   // 奖励百分比(如35.00表示35%)
     private Date createdAt;      // 记录创建时间
     private Date updatedAt;      // 记录更新时间
 
@@ -34,11 +35,11 @@ public class RewardRulesVo {
         this.rankOrder = rankOrder;
     }
 
-    public Double getPercentage() {
+    public BigDecimal getPercentage() {
         return percentage;
     }
 
-    public void setPercentage(Double percentage) {
+    public void setPercentage(BigDecimal percentage) {
         this.percentage = percentage;
     }
